Check if admin account credit service is available
Get credit account balance before granting an offer and do not proceed
if that fails. This does not protect against cases where the admin
account credit service goes down after the offer is granted, but it is
still beneficial in scenarios where the admin account credit service is
already down before the event processing starts.